fix: service moved to /service
