+++ src/app/services/icon-registry.service.ts
